reusable base class for kapua service application
