+++ service-app/job-engine/src/main/java/org/eclipse/kapua/service/job/engine/app/JobEngineApplication.java
@@ -11,27 +11,11 @@
  *******************************************************************************/
 package org.eclipse.kapua.service.job.engine.app;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.URL;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
-import org.apache.shiro.SecurityUtils;
-import org.apache.shiro.config.Ini;
-import org.apache.shiro.config.IniSecurityManagerFactory;
-import org.apache.shiro.mgt.SecurityManager;
-import org.eclipse.kapua.KapuaErrorCodes;
-import org.eclipse.kapua.KapuaException;
-import org.eclipse.kapua.commons.jpa.JdbcConnectionUrlResolvers;
-import org.eclipse.kapua.commons.liquibase.KapuaLiquibaseClient;
-import org.eclipse.kapua.commons.setting.system.SystemSetting;
-import org.eclipse.kapua.commons.setting.system.SystemSettingKey;
-import org.eclipse.kapua.commons.util.KapuaFileUtils;
-import org.eclipse.kapua.commons.util.xml.XmlUtil;
-import org.eclipse.kapua.service.commons.app.ApplicationBase;
 import org.eclipse.kapua.service.commons.app.Context;
+import org.eclipse.kapua.service.commons.app.KapuaServiceApplication;
 import org.eclipse.kapua.service.commons.http.HttpServiceContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -40,14 +24,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.liquibase.LiquibaseServiceLocatorApplicationListener;
 
-import com.fasterxml.jackson.databind.Module;
-import com.google.common.base.MoreObjects;
-
-import io.vertx.core.Future;
-import io.vertx.core.json.Json;
-
 @SpringBootApplication
-public class JobEngineApplication extends ApplicationBase<JobEngineApplicationConfiguration> {
+public class JobEngineApplication extends KapuaServiceApplication<JobEngineApplicationConfiguration> {
 
     Logger logger = LoggerFactory.getLogger(JobEngineApplication.class);
 
@@ -63,65 +41,11 @@ public class JobEngineApplication extends ApplicationBase<JobEngineApplicationCo
     }
 
     @Override
-    protected void runInternal(Context context, JobEngineApplicationConfiguration config, Future<Void> runFuture) throws Exception {
+    protected void doContext(Context context, JobEngineApplicationConfiguration config) throws Exception {
         Objects.requireNonNull(context, "param: context");
         Objects.requireNonNull(config, "param: config");
-        Objects.requireNonNull(runFuture, "param: runFuture");
 
-        context.getVertx().executeBlocking(runExecution -> {
-
-            try {
-                SystemSetting sysConfig = SystemSetting.getInstance();
-                if (sysConfig.getBoolean(SystemSettingKey.DB_SCHEMA_UPDATE, false)) {
-                    logger.info("Initialize Liquibase embedded client.");
-                    String dbUsername = sysConfig.getString(SystemSettingKey.DB_USERNAME);
-                    String dbPassword = sysConfig.getString(SystemSettingKey.DB_PASSWORD);
-                    String schema = MoreObjects.firstNonNull(sysConfig.getString(SystemSettingKey.DB_SCHEMA_ENV), sysConfig.getString(SystemSettingKey.DB_SCHEMA));
-
-                    // initialize driver
-                    Class.forName(sysConfig.getString(SystemSettingKey.DB_JDBC_DRIVER));
-                    logger.debug("Starting Liquibase embedded client update - URL: {}, user/pass: {}/{}", JdbcConnectionUrlResolvers.resolveJdbcUrl(), dbUsername, dbPassword);
-                    new KapuaLiquibaseClient(JdbcConnectionUrlResolvers.resolveJdbcUrl(), dbUsername, dbPassword, Optional.of(schema)).update();
-                }
-
-                // Shiro, JAXB and Jackson initialization
-                initShiro();
-                XmlUtil.setContextProvider(new JobEngineJaxbContextProvider());
-                registerJacksonModule(new JobEngineModule());
-
-                // Configure Services
-                context.getServiceContext("jobEngineService", HttpServiceContext.class).addController(config.getJobEngineHttpController());
-                runExecution.complete();
-            } catch (Exception exc) {
-                runExecution.fail(exc);
-            }
-
-        }, resultHandler -> {
-            if (resultHandler.succeeded()) {
-                runFuture.complete();
-            } else {
-                runFuture.fail(resultHandler.cause());
-            }
-        });
-    }
-
-    private void registerJacksonModule(Module jacksonModule) {
-        Json.mapper.registerModule(jacksonModule);
-        Json.prettyMapper.registerModule(jacksonModule);
-    }
-
-    private void initShiro() throws KapuaException {
-        // initialize shiro context for broker plugin from shiro ini file
-        final URL shiroIniUrl = KapuaFileUtils.getAsURL("shiro.ini");
-        Ini shiroIni = new Ini();
-        try (final InputStream input = shiroIniUrl.openStream()) {
-            shiroIni.load(input);
-        } catch (IOException ex) {
-            logger.error("Error loading shiro.ini", ex);
-            throw new KapuaException(KapuaErrorCodes.INTERNAL_ERROR);
-        }
-
-        SecurityManager securityManager = new IniSecurityManagerFactory(shiroIni).getInstance();
-        SecurityUtils.setSecurityManager(securityManager);
+        // Configure Services
+        context.getServiceContext("jobEngineService", HttpServiceContext.class).addController(config.getJobEngineHttpController());
     }
 }
