+++ service-app/job-engine/src/main/resources/application.properties
@@ -16,7 +16,6 @@ monitoring.endpoint.securePort=9444
 monitoring.endpoint.tls=true
 
 # Services
-services.configs.jobEngineService.name=jobEngineService
 services.configs.jobEngineService.type=http
 services.configs.jobEngineService.instances=1
 services.configs.jobEngineService.properties.basePath=v1
